python_services_archive/python/requirements.txt
Flask==2.3.3
python-docx==0.8.11
reportlab==4.0.4
docx2pdf==0.1.8
